requirements.txt
-e git+https://github.com/Eder175/a121_backend.git@5c89e0f9e0c77b963850ab5d60eb91a1d79abf3c#egg=a121_backend
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
backports.tarfile==1.2.0
billiard==4.2.1
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
django-meta==2.4.2
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.3
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-dialogflow==2.41.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
kombu==5.5.0
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
pbs-installer==2025.3.11
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
RapidFuzz==3.12.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.3.13.13
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.2
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
zstandard==0.23.0
